+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
     - BUGFIX: fix HLS/m3u8 based recording by using a more robust and correct timing
     - BUGFIX: make HLS/m3u8 based playing more robust by using HLS.js API more correct
     - IMPROVEMENT: optically improve recording list on login dialog
+    - IMPROVEMENT: add "REPLAY" information overlay to video playback
 
 - 1.8.6
     - IMPROVEMENT: use my github.oscdn.org CDN service to circumvent Amazon S3 peering bottlenecks